requirements.txt
# FastAPI and web framework
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9

# Configuration and environment
python-dotenv==1.0.1

# Data processing and ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2

# AI/ML explainability and fairness
aif360==0.6.1
shap==0.44.0
lime==0.2.0.1

# Model tracking and versioning
mlflow==2.11.2

# Database
duckdb==0.10.2

# HTTP client and API requests
httpx==0.27.2
requests==2.32.3

# Data validation and models
pydantic==2.11.4

# PDF generation for reports
fpdf2==2.8.1

# Development and testing (optional)
pytest==8.3.3
pytest-asyncio==0.24.0